Split authorization middleware, it was actually doing two things: authorize and get database connection params
