+++ app/middlewares/authorization.js
@@ -1,7 +1,7 @@
 const AuthApi = require('../auth/auth_api');
 const basicAuth = require('basic-auth');
 
-module.exports = function authorization (metadataBackend, userDatabaseService, forceToBeAuthenticated = false) {
+module.exports = function authorization (metadataBackend, forceToBeAuthenticated = false) {
     return function authorizationMiddleware (req, res, next) {
         const { user } = res.locals;
         const credentials = getCredentialsFromRequest(req);
@@ -16,6 +16,10 @@ module.exports = function authorization (metadataBackend, userDatabaseService, f
         const authApi = new AuthApi(req, params);
 
         authApi.verifyCredentials({}, function (err, authenticated) {
+            if (req.profiler) {
+                req.profiler.done('verifyCredentials');
+            }
+
             if (err) {
                 return next(err);
             }
@@ -26,24 +30,7 @@ module.exports = function authorization (metadataBackend, userDatabaseService, f
                 return next(new Error('permission denied'));
             }
 
-            const apikeyToken = res.locals.api_key;
-
-            userDatabaseService.getConnectionParams(user, apikeyToken, authenticated,
-                function (err, userDbParams, authDbParams, userLimits) {
-                if (req.profiler) {
-                    req.profiler.done('setDBAuth');
-                }
-
-                if (err) {
-                    return next(err);
-                }
-
-                res.locals.userDbParams = userDbParams;
-                res.locals.authDbParams = authDbParams;
-                res.locals.userLimits = userLimits;
-
-                next();
-            });
+            next();
         });
     };
 };
